:sparkles: feat: Add table of contents store and cache logic
This commit adds a new file src/stores/table-contents.js that defines a
pinia store for the table of contents data. It also modifies the
src/api/chapters.js file to use the pinia store as a cache layer and
accept an onCache callback. Finally, it updates the src/pages/info.vue
file to use the onCache callback and set the loading state accordingly.